% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimates.R
\name{estimate}
\alias{estimate}
\title{Estimate Self-Specified Target Using Mixed Residual Bootstrap Samples}
\usage{
estimate(X, y, nboot, DefFunction)
}
\arguments{
\item{X}{The design matrix with dimension n by p, where n represents the number of observations and p represents the number of parameters.}

\item{y}{The n observations of the dependent variable y.}

\item{nboot}{Specify the number of Mixed Residual Bootstrap samples that you want to generate. Noted that the input should be a positive integar.}

\item{DefFunction}{It should be a function of X or y. The first input of DefFunction should be the bootstrapping X and the second input should be the bootstrapping y. Specifically, you may write DefFunction <- function(X_star,y_star){...}}
}
\value{
Estimate It is a bootstrap estimate of your specified target. Estimate is a list and your estimated target can be a single value or a vector.

se It is the standard error estimated by the bootstrap samples.
}
\description{
This function will automatically generate the estimates of user-specified statistics using the mixed residual bootstrap method.
}
\examples{
\dontrun{
p <- 30
n <- 100
X <- matrix(rnorm(p*n),ncol = p,nrow = n)
beta <- as.vector(rnorm(p,10,4))
y <- X\%*\%beta+rnorm(n)
fun <- function(X,y){
 f <- (t(X)\%*\%X)\%*\%t(X)\%*\%y
 return(f)
}
estimates <- estimate(X,y,100,fun)
}
}
\references{
Lee, S. M. S. & Wu Y.L. (2017). A Bootstrap Recipe for Post-Model-Selection Inference
}
\seealso{
\code{\link[base]{paste}}
}
