% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateSample.R
\name{generate}
\alias{generate}
\title{Generate Mixed Residual Bootstrap Samples}
\usage{
generate(name)
}
\arguments{
\item{name}{a string that will appear before "Hello, world!"}
}
\description{
It is a very useful function to generate Mixed Residual Bootstrap Samples.
}
\examples{
\dontrun{
hello("Linda")
[1] "Linda Hello, world!"
hello("Adam")
[1] "Adam Hello, world!"
}
}
\seealso{
\code{\link[base]{paste}}
}
