% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inference.R
\name{inference}
\alias{inference}
\title{Statistical Inference Using Mixed Residual Bootstrap Samples}
\usage{
inference(X, y, nboot, DefFunction, alpha, plotoption)
}
\arguments{
\item{X}{The design matrix with dimension n by p, where n represents the number of observations and p represents the number of parameters.}

\item{y}{The n observations of the dependent variable y.}

\item{nboot}{Specify the number of Mixed Residual Bootstrap samples that you want to generate. Noted that the input should be a positive integar.}

\item{DefFunction}{It should be a function of X or y. The first input of DefFunction should be the bootstrapping X and the second input should be the bootstrapping y. Specifically, the output of the function must be a single value.}

\item{alpha}{The significance level. It should be a value between 0 and 1.}

\item{plotoption}{The default setting is TRUE. A density plot will be generated automatically. If you want to eliminate this option, please use plot=FALSE.}
}
\value{
lwb It is the lower bound of the confidence interval.

upb It is the upper bound of the confidence interval.

A distribution plot will be generated automatically.
}
\description{
This function will automatically provide the inference results using the mixed residual bootstrap method.
}
\examples{
\dontrun{
p <- 30
n <- 100
X <- matrix(rnorm(p*n),ncol = p,nrow = n)
beta <- as.vector(rnorm(p,10,4))
y <- X\%*\%beta+rnorm(n)
fun <- function(X,y){
 f <- sum(solve(t(X)\%*\%X)\%*\%t(X)\%*\%y)
 return(f)
}
tt <- inference(X,y,100,fun,0.05,TRUE)
lwb <- tt$lwb
upb <- tt$upb
}
}
\references{
Lee, S. M. S. & Wu Y.L. (2017). A Bootstrap Recipe for Post-Model-Selection Inference
}
