% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rm_abbreviation.R
\docType{data}
\name{rm_abbreviation}
\alias{rm_abbreviation}
\title{Abbreviations}
\format{\preformatted{Classes 'regexr', 'character'  atomic [1:1] ([A-Za-z][\\.]\\s*)\{1,\}([A-Za-z][\\.])
  ..- attr(*, "subs")=List of 2
  .. ..$ let_per_1:Classes 'subcom', 'character'  atomic [1:1] ([A-Za-z][\\.]\\s*)\{1,\}
  .. .. .. ..- attr(*, "comment")= chr "Letter folowed by period and optional space (1 or more times)"
  .. ..$ let_per_2:Classes 'subcom', 'character'  atomic [1:1] ([A-Za-z][\\.])
  .. .. .. ..- attr(*, "comment")= chr "Ending letter folowed by period"
  ..- attr(*, "comments")=List of 2
  .. ..$ let_per_1: chr "Letter folowed by period and optional space (1 or more times)"
  .. ..$ let_per_2: chr "Ending letter folowed by period"
}}
\usage{
rm_abbreviation
}
\description{
Find abbreviates that contain capitals or or lower case.  Must have at
minimum 1 letter followed by a period folowed by another letter and period.
May contain additional letters and spaces.
}
\section{Regex}{
 TRUE
}
\examples{
input <- c("I want $2.33 at 2:30 p.m. to go to A.n.p.",
    "She will send it A.S.A.P. (e.g. as soon as you can) said I.",
    "Hello world.", "In the U. S. A.")

regmatches(input, gregexpr(rm_abbreviation, input, perl = TRUE))
gsub(rm_abbreviation, "", input, perl = TRUE)
strsplit(input, rm_abbreviation, perl = TRUE)
}
\keyword{datasets}

