% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunPerm.R
\name{RunPerm}
\alias{RunPerm}
\title{Run permutations}
\usage{
RunPerm(input, replicates)
}
\arguments{
\item{input}{The input dataframe (a dataframe object)}

\item{replicates}{Number of permutation replicates to perform (an integer)}
}
\value{
A dataframe.
}
\description{
Run permutations
}
\examples{

data <- read.table("Example_file.txt")
RunPerm(input = data,replicates = 50)

}
