% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimates.R
\name{estimate}
\alias{estimate}
\title{Estimate Self-Specified Target Using Mixed Residual Bootstrap Samples}
\usage{
estimate(X, y, nboot, DefFunction)
}
\arguments{
\item{X}{The design matrix with dimension n by p, where n represents the number of observations and p represents the number of parameters.}

\item{y}{The n observations of the dependent variable y.}

\item{nboot}{Specify the number of Mixed Residual Bootstrap samples that you want to generate. Noted that the input should be a positive integar.}

\item{DefFunction}{It should be a function of X or y. The first input of DefFunction should be the bootstrapping X and the second input should be the bootstrapping y. Specifically, you may write DefFunction <- function(X_star,y_star){...}}
}
\value{
Estimate It is a bootstrap estimate of your specified target.

se It is the standard error estimated by the bootstrap samples.
}
\description{
This function will automatically generate the estimates of user-specified statistics using the mixed residual bootstrap method.
}
\examples{
\dontrun{
hello("Linda")
[1] "Linda Hello, world!"
hello("Adam")
[1] "Adam Hello, world!"
}
}
\references{
Lee, S. M. S. & Wu Y.L. (2017). A Bootstrap Recipe for Post-Model-Selection Inference
}
\seealso{
\code{\link[base]{paste}}
}
