% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Full_script.R
\name{Full_script}
\alias{Full_script}
\title{Running all the steps of the SAMPLE pipeline}
\usage{
Full_script(
  input,
  output_N = "Results",
  replicates = 50,
  stability_thres = 2,
  sucess_points = 10,
  diff = 1
)
}
\arguments{
\item{input}{Input dataframe (a dataframe object)}

\item{output_N}{Prefix used for the output (a character; default output_N="Results")}

\item{replicates}{Nnumber of permutation replicates to perform (an integer; default replicates=50)}

\item{stability_thres}{Threshold used to define stability (an integer). This value will be divided by the square root of the number of replicates (a float, default stability_thres = 2.0).}

\item{sucess_points}{Number of successive (mean) prevalence rates that are below a threshold (see parameter stability_thres) used to define stability (an integer; default stability_thres = 10).}

\item{diff}{Difference between the absolute minimum and maximum values among the all the means used to set the stability threshold (a float, default diff = 1.0).}
}
\value{
A textfile and a pdf.
}
\description{
Running all the steps of the SAMPLE pipeline
}
\examples{
data("coral_symbionts")
set.seed(812)
Full_script(input = coral_symbionts,output_N = "Example",replicates = 50,stability_thres = 2,sucess_points = 10,diff = 1)
}
