% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateSample.R
\name{generate}
\alias{generate}
\title{Generate Mixed Residual Bootstrap Samples}
\usage{
generate(X, y, nboot)
}
\arguments{
\item{X}{The design matrix with dimension n by p, where n represents the number of observations and p represents the number of parameters.}

\item{y}{The n observations of the dependent variable y.}

\item{nboot}{Specify the number of Mixed Residual Bootstrap samples that you want to generate. Noted that the input should be a positive integar.}
}
\value{
returnlist
}
\description{
It is a very useful function to generate Mixed Residual Bootstrap Samples.
}
\examples{
\dontrun{
hello("Linda")
[1] "Linda Hello, world!"
hello("Adam")
[1] "Adam Hello, world!"
}
}
\seealso{
\code{\link[base]{paste}}
}
