% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateSample.R
\name{generate}
\alias{generate}
\title{Generate Mixed Residual Bootstrap Samples}
\usage{
generate(X, y, nboot)
}
\arguments{
\item{X}{The design matrix with dimension n by p, where n represents the number of observations and p represents the number of parameters.}

\item{y}{The n observations of the dependent variable y.}

\item{nboot}{Specify the number of Mixed Residual Bootstrap samples that you want to generate. Noted that the input should be a positive integar.}
}
\value{
X   A list of nboot bootstrapping design matrices

y   A list of nboot bootstrapping dependend variable observations
}
\description{
It is a very useful function to generate Mixed Residual Bootstrap Samples. The default setting of model selection procedure in this version is 10-fold Cross-Validation LASSO. We will update the package to include more procedures later.
}
\examples{
\dontrun{
p <- 30
n <- 100
X <- Matrix(rnorm(p*n),ncol = p,nrow = n)
beta <- as.vector(rnorm(p,10,4))
y <- X\%*\%beta+rnorm(n)
BootResult <- generate(X,y,100)
X_star <- BootResult$X
y_star <- BootResult$y
}
}
\references{
Lee, S. M. S. & Wu Y.L. (2017). A Bootstrap Recipe for Post-Model-Selection Inference
}
\seealso{
\code{\link[base]{paste}}
}
