% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotstab.R
\name{plotstab}
\alias{plotstab}
\title{Plot}
\usage{
plotstab(data, info, outputName = "Results", outputDir)
}
\arguments{
\item{data}{A dataframe generated by the function RunPerm() (a dataframe object)}

\item{info}{A dataframe generated by the function StabAssess() (a dataframe object)}

\item{outputName}{Prefix used for the output (a character; default output_N="Results")}

\item{outputDir}{Set output directory (a character)}
}
\value{
A pdf and a text file
}
\description{
Plot
}
\examples{

data("coral_symbionts")
set.seed(812)
perm = RunPerm(input = coral_symbionts,replicates = 50)
stability = StabAssess(data = perm,stability_thres = 5 ,sucess_points = 5,diff = 2 )
plotstab(data = perm, info = stability, outputName = "Stability_example", outputDir = "./")

}
