% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotstab.R
\name{plotstab}
\alias{plotstab}
\title{Plot}
\usage{
plotstab(data, info, output_N = "Results", outputDir)
}
\arguments{
\item{data}{A dataframe generated by the function RunPerm() (a dataframe object)}

\item{info}{A dataframe generated by the function StabAssess() (a dataframe object)}

\item{output_N}{Prefix used for the output (a character; default output_N="Results")}

\item{outputDir}{Set output directory (a character)}
}
\description{
Plot
}
\examples{

data <- read.table("Example_file.txt")
perm = RunPerm(input = data,replicates = 50)
stability = StabAssess(data = perm,stability_thres = 5 ,sucess_points = 5,diff = 2 )
plotstab(data = perm, info = stability, outputName = "Example", outputDir = "./")

}
