% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunPerm.R
\name{RunPerm}
\alias{RunPerm}
\title{Run permutations}
\usage{
RunPerm(input, replicates = 50)
}
\arguments{
\item{input}{The input dataframe (a dataframe object)}

\item{replicates}{Number of permutation replicates to perform (an integer; default replicates=50)}
}
\value{
A dataframe.
}
\description{
Run permutations
}
\examples{
data("coral_symbionts")
RunPerm(input = coral_symbionts,replicates = 50)

}
